fix: refactor code to seperate components
